+++ integration_test/cases/preload.exs
@@ -1,7 +1,7 @@
 defmodule Ecto.Integration.PreloadTest do
   use Ecto.Integration.Case
 
-  require Ecto.Integration.TestRepo, as: TestRepo
+  alias Ecto.Integration.TestRepo
   import Ecto.Query
 
   alias Ecto.Integration.Post