+++ integration_test/sql/transaction.exs
@@ -4,8 +4,8 @@ defmodule Ecto.Integration.TransactionTest do
   use ExUnit.Case, async: true
 
   import Ecto.Query
-  require Ecto.Integration.PoolRepo, as: PoolRepo
-  require Ecto.Integration.TestRepo, as: TestRepo
+  alias Ecto.Integration.PoolRepo
+  alias Ecto.Integration.TestRepo
 
   defmodule UniqueError do
     defexception [:message]