+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.14.0-dev
 
 * Enhancements
+  * Add `Ecto.Adapters.SQL.to_sql/3`
   * Store the connection PID in `Ecto.LogEntry`
   * Add `Ecto.Pools.SojournBroker` as a more flexible and customizable alternative to `Ecto.Pools.Poolboy`
   * Support `:on_delete` when defining `has_many` and `belongs_to` in schema